refactor(plugin-assets:init): separates init actions from command
